Move mau-extratests-phub out of common
We need to remove a module from a single version. The only way to do that is
to explicitly add all the other versions to the conditional schedule.
To avoid that I created a yaml in each directory.

Additionally we had forgotten to add two modules for 15SP4 which are now added.

Relevant tickets:
https://progress.opensuse.org/issues/113177
https://progress.opensuse.org/issues/113216